We added an int function that runs a winch motor to raise and lower the climber.

+++ src/main/cpp/Climber.cpp
@@ -11,7 +11,7 @@ Climber::Climber(Robot * pRobot)
 
 
 
-void Climber::RunWinch(float speed)
+void Climber::RunWinch(int speed)
 {
     WinchMotor.Set(speed);
 }
